+++ src/components/InfiniteScroll/Scroller.stories.tsx
@@ -4,7 +4,7 @@ import faker from 'faker';
 
 // There is a problem using useEndpoint from stories?
 import useEndpoint from 'hooks/useEndpoint';
-import Component from './InfiniteScroll';
+import Component from './temp';
 
 const createNames = (num: number) => {
     const list = [];
@@ -21,40 +21,42 @@ function getNames() {
 }
 
 export default { title: 'Components/InfiniteScroll' };
+
 export function ExampleScroll() {
-    // const [state, setState] = React.useState(2000);
-    const [names, setNames] = React.useState<string[]>(createNames(50));
-    const [hasNext, setHasNext] = React.useState(true);
+    return <Component />;
+    // // const [state, setState] = React.useState(2000);
+    // const [names, setNames] = React.useState<string[]>(createNames(50));
+    // const [hasNext, setHasNext] = React.useState(true);
 
-    const apiRequest = React.useCallback(() => getNames(), []);
+    // const apiRequest = React.useCallback(() => getNames(), []);
 
-    const [sendRequest, isLoading] = useEndpoint(apiRequest, {
-        onSuccess: (results) => {
-            console.log('Success');
-            setNames((prev) => [...prev, ...results.data.names]);
-            setHasNext(results.data.hasNext);
-        },
-        onFailure: () => console.log('Failure'),
-    });
+    // const [sendRequest, isLoading] = useEndpoint(apiRequest, {
+    //     onSuccess: (results) => {
+    //         console.log('Success');
+    //         setNames((prev) => [...prev, ...results.data.names]);
+    //         setHasNext(results.data.hasNext);
+    //     },
+    //     onFailure: () => console.log('Failure'),
+    // });
 
-    // const addNames = () => {
-    //     console.log('Waiting');
-    //     setIsLoading(true);
-    //     setTimeout(() => {
-    //         console.log('Finished waiting');
-    //         setIsLoading(false);
-    //         setNames((prev) => [...prev, ...createNames(30)]);
-    //     }, 5000);
-    // };
-    return (
-        <Component
-            loadMore={hasNext ? sendRequest : null}
-            isLoading={isLoading}
-        >
-            {names.map((name, index) => (
-                <h1 key={index}>{name}</h1>
-            ))}
-            {/* <div style={{ height: `${state}px` }} /> */}
-        </Component>
-    );
+    // // const addNames = () => {
+    // //     console.log('Waiting');
+    // //     setIsLoading(true);
+    // //     setTimeout(() => {
+    // //         console.log('Finished waiting');
+    // //         setIsLoading(false);
+    // //         setNames((prev) => [...prev, ...createNames(30)]);
+    // //     }, 5000);
+    // // };
+    // return (
+    //     <Component
+    //         loadMore={hasNext ? sendRequest : null}
+    //         isLoading={isLoading}
+    //     >
+    //         {names.map((name, index) => (
+    //             <h1 key={index}>{name}</h1>
+    //         ))}
+    //         {/* <div style={{ height: `${state}px` }} /> */}
+    //     </Component>
+    // );
 }
